services/auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6

